feat: apply-join AddColumn should only calculate offsets when needed
