+++ go/vt/vtgate/planbuilder/operators/filter.go
@@ -113,3 +113,27 @@ func (f *Filter) Compact(*plancontext.PlanningContext) (ops.Operator, rewrite.Ap
 	f.Predicates = append(f.Predicates, other.Predicates...)
 	return f, rewrite.NewTree, nil
 }
+
+func (f *Filter) planOffsets(ctx *plancontext.PlanningContext) error {
+	resolveColumn := func(col *sqlparser.ColName) (int, error) {
+		newSrc, offset, err := f.Source.AddColumn(ctx, aeWrap(col))
+		if err != nil {
+			return 0, err
+		}
+		f.Source = newSrc
+		return offset, nil
+	}
+	cfg := &evalengine.Config{
+		ResolveType:   ctx.SemTable.TypeForExpr,
+		Collation:     ctx.SemTable.Collation,
+		ResolveColumn: resolveColumn,
+	}
+
+	eexpr, err := evalengine.Translate(sqlparser.AndExpressions(f.Predicates...), cfg)
+	if err != nil {
+		return err
+	}
+
+	f.FinalPredicate = eexpr
+	return nil
+}
