+++ go/vt/vtgate/planbuilder/testdata/from_cases.json
@@ -942,14 +942,14 @@
         "Variant": "LeftJoin",
         "JoinColumnIndexes": "L:0,R:0",
         "JoinVars": {
-          "m1_col": 0
+          "m1_col": 1
         },
         "TableName": "`user`_unsharded_unsharded",
         "Inputs": [
           {
             "OperatorType": "Join",
             "Variant": "LeftJoin",
-            "JoinColumnIndexes": "R:0,L:0",
+            "JoinColumnIndexes": "L:0,R:0",
             "JoinVars": {
               "user_col": 0
             },
@@ -5548,9 +5548,9 @@
               {
                 "OperatorType": "Join",
                 "Variant": "LeftJoin",
-                "JoinColumnIndexes": "L:1,R:0",
+                "JoinColumnIndexes": "L:0,R:0",
                 "JoinVars": {
-                  "user_col": 0
+                  "user_col": 1
                 },
                 "TableName": "`user`_user_extra",
                 "Inputs": [
@@ -5561,8 +5561,8 @@
                       "Name": "user",
                       "Sharded": true
                     },
-                    "FieldQuery": "select `user`.col, `user`.id from `user` where 1 != 1",
-                    "Query": "select `user`.col, `user`.id from `user`",
+                    "FieldQuery": "select `user`.id, `user`.col from `user` where 1 != 1",
+                    "Query": "select `user`.id, `user`.col from `user`",
                     "Table": "`user`"
                   },
                   {
@@ -6304,9 +6304,9 @@
               {
                 "OperatorType": "Join",
                 "Variant": "LeftJoin",
-                "JoinColumnIndexes": "L:1,R:0",
+                "JoinColumnIndexes": "L:0,R:0",
                 "JoinVars": {
-                  "user_col": 0
+                  "user_col": 1
                 },
                 "TableName": "`user`_user_extra",
                 "Inputs": [
@@ -6317,8 +6317,8 @@
                       "Name": "user",
                       "Sharded": true
                     },
-                    "FieldQuery": "select `user`.col, `user`.id from `user` where 1 != 1",
-                    "Query": "select `user`.col, `user`.id from `user`",
+                    "FieldQuery": "select `user`.id, `user`.col from `user` where 1 != 1",
+                    "Query": "select `user`.id, `user`.col from `user`",
                     "Table": "`user`"
                   },
                   {
